requirements.txt
absl-py==0.7.1
astor==0.8.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
cycler==0.10.0
Cython==0.29.10
dominate==2.3.5
Flask==1.0.3
flask-appconfig==0.11.1
Flask-Bootstrap==3.3.7.1
Flask-Debug==0.4.3
flask-nav==0.6
Flask-WTF==0.14.2
gast==0.2.2
google-pasta==0.1.7
grpcio==1.21.1
h5py==2.9.0
idna==2.8
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.0
numpy==1.16.4
object-detection==0.1
opencv-python==4.1.0.25
pickleshare==0.7.5
Pillow==6.0.0
protobuf==3.8.0
pyparsing==2.4.0
python-dateutil==2.8.0
requests==2.22.0
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
sklearn==0.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
urllib3==1.25.3
visitor==0.1.3
Werkzeug==0.15.4
wrapt==1.11.2
WTForms==2.2.1
